Incorporate various TUV patches: r1836,r1837,r1838,r1839,r1840,r1841,r1842,r1843,r1849,r1850,r1859 (various code clean up, compile fixes, Windows fixes, new UriParameter class)
git-svn-id: https://svn.libreccm.org/ccm/trunk@98 8810af33-2d31-482b-a856-94f89814c4df

+++ ccm-core/test/src/com/arsdigita/categorization/CategoryTestCase.java
@@ -198,7 +198,7 @@ public class CategoryTestCase extends BaseTestCase {
                     }
                 } else if (currentValue instanceof BigInteger) {
                     test = ((BigInteger)currentValue)
-                        .compareTo((BigDecimal)prevValue) >= 0;
+                        .compareTo((BigInteger)prevValue) >= 0;
                 } else if (currentValue instanceof Number) {
                     test = ((Number)currentValue).longValue() >=
                         ((Number)prevValue).longValue();
